requirements.txt
uvicorn
sse-starlette
fastapi
sseclient
fastapi-utils
aiohttp
python-dotenv
httpx
pytest
pytest-asyncio
